requirements-linux.lock.txt
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
Flask==3.0.3
Flask-JWT-Extended==4.7.1
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.4.0
multidict==6.6.4
openai==1.99.9
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.11.2
packaging==25.0
pipx==1.7.1
platformdirs==4.3.8
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.7.3
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.6.0
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
